Plot timeseries fill
Function plots two time series and allows the user to fill the area between each point in time within the plot. This is still under development. The geom_ribbon does not seem to work for cross over interactions, nor did creating a series of polygons or area plots. But there is likely a better way to do this to get complete fill between the two, this was just the one that looked the cleanest.