Regression and ensembling with tidymodels
Contains workflow functions for performing regression (OLS, ridge, lasso, and elastic) models. User has the ability to select ensembling methods or can choose to run models individually to compare results directly. This is the first version of the functions, but additional naming and visualization will be added for clarity and enhancing the workflow.